Move UUID into it's own package
